+++ components/typography/styles.js
@@ -1,23 +1,7 @@
 import '../colors/colors.js';
 import { css, unsafeCSS } from 'lit';
+import { _isValidCssSelector } from '../../helpers/internal/css.js';
 import { svgToCSS } from '../../helpers/svg-to-css.js';
-
-export const _isValidCssSelector = (selector) => {
-	const partIsValid = (part) => {
-		const re = /([a-zA-Z0-9-_ >.#]+)(\[[a-zA-Z0-9-_]+\])?([a-zA-Z0-9-_ >.#]+)?/g;
-		if (part === ':host') return true;
-		const match = part.match(re);
-		const isValid = !!match && match.length === 1 && match[0].length === part.length;
-		if (!isValid) {
-			console.warn(`Invalid CSS selector: "${part}"`);
-		}
-		return isValid;
-	};
-
-	const parts = selector.split(',');
-	const allValid = parts.every(part => partIsValid(part));
-	return allValid;
-};
 
 /**
  * A private helper method that should not be used by general consumers
